pyproject.toml
[project]
name = "rate-limit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "typing-extensions>=4.15.0",
]
